// src/app/model/project.model.ts
export class Project {
  // name: string;
  // description: string;
  // techStack: string[];
  // frontendSourceCode: string;
  // backendSourceCode: string;
  name: string = '';
  description: string = '';
  techStack: string[] = ['asddsd'];
  frontendSourceCode: string = '';
  backendSourceCode: string = '';
  imagePath: string = '';

  // constructor(name: string,
  //             description: string,
  //             techStack: string[],
  //             frontendSourceCode: string,
  //             backendSourceCode: string) {
  //   this.name = name;
  //   this.description = description;
  //   this.techStack = techStack;
  //   this.frontendSourceCode = frontendSourceCode;
  //   this.backendSourceCode = backendSourceCode;
  // }
}